fix: move usePlatformApi ldFlag from personal card service to page
